croppable link parser
