+++ infra/link/vmess3p/common.go
@@ -1,0 +1,20 @@
+package vmess3p
+
+import (
+	"encoding/base64"
+	"strings"
+)
+
+func base64Decode(b64 string) ([]byte, error) {
+	b64 = strings.TrimSpace(b64)
+	stdb64 := b64
+	if pad := len(b64) % 4; pad != 0 {
+		stdb64 += strings.Repeat("=", 4-pad)
+	}
+
+	b, err := base64.StdEncoding.DecodeString(stdb64)
+	if err != nil {
+		return base64.URLEncoding.DecodeString(b64)
+	}
+	return b, nil
+}
