+++ infra/link/vmess3p/vmess3p.go
@@ -1,4 +1,4 @@
-package link
+package vmess3p
 
 import (
 	"encoding/json"
@@ -8,8 +8,39 @@ import (
 	"v2ray.com/core/infra/conf"
 )
 
-// ToOutbound converts the vmess link to *OutboundDetourConfig
-func (v *VMessLink) ToOutbound() *conf.OutboundDetourConfig {
+// TPLink represents a third party vmess link
+type TPLink struct {
+	Ver      string      `json:"v,omitempty"`
+	Add      string      `json:"add,omitempty"`
+	Aid      interface{} `json:"aid,omitempty"`
+	Host     string      `json:"host,omitempty"`
+	ID       string      `json:"id,omitempty"`
+	Net      string      `json:"net,omitempty"`
+	Path     string      `json:"path,omitempty"`
+	Port     interface{} `json:"port,omitempty"`
+	Ps       string      `json:"ps,omitempty"`
+	TLS      string      `json:"tls,omitempty"`
+	Type     string      `json:"type,omitempty"`
+	OrigLink string      `json:"-,omitempty"`
+}
+
+// ToString implements Link.ToString()
+func (v TPLink) ToString() string {
+	return v.OrigLink
+}
+
+// Tag implements Link.Tag()
+func (v *TPLink) Tag() string {
+	return v.Ps
+}
+
+// Detail implements Link.Detail()
+func (v TPLink) Detail() string {
+	return fmt.Sprintf("Net: %s\nAddr: %s\nPort: %v\nUUID: %s\nType: %s\nTLS: %s\nPS: %s\n", v.Net, v.Add, v.Port, v.ID, v.Type, v.TLS, v.Ps)
+}
+
+// ToOutbound implements Link.ToOutbound()
+func (v *TPLink) ToOutbound() *conf.OutboundDetourConfig {
 	out := &conf.OutboundDetourConfig{}
 	out.Protocol = "vmess"
 
@@ -85,4 +116,23 @@ func (v *VMessLink) ToOutbound() *conf.OutboundDetourConfig {
 }`, v.Add, v.Port, v.ID, v.Aid)))
 	out.Settings = &oset
 	return out
+}
+
+// IsEqual tests if this vmess link is equal to another
+func (v *TPLink) IsEqual(c *TPLink) bool {
+	realNet := func(n string) string {
+		if n == "" {
+			return "tcp"
+		}
+		return n
+	}
+	if realNet(v.Net) != realNet(c.Net) {
+		return false
+	}
+	if fmt.Sprintf("%v", c.Port) != fmt.Sprintf("%v", v.Port) {
+		return false
+	}
+
+	return v.Add == c.Add && v.Aid == c.Aid && v.Host == c.Host && v.ID == c.ID &&
+		v.Path == c.Path && v.TLS == c.TLS && v.Type == c.Type
 }
