+++ main/distro/all/all.go
@@ -66,4 +66,7 @@ import (
 
 	// commands
 	_ "v2ray.com/core/main/commands/all"
+
+	// third party vmess link support
+	_ "v2ray.com/core/infra/link/vmess3p"
 )
